+++ src/ruby/models/music.rb
@@ -18,19 +18,6 @@ class Music
     }
   end
 
-  # Set the volume.
-  # @param value [Float] A value between 0 and 1.
-  # @raise [ArgumentError] If the value is out of bounds.
-  # @return [nil]
-  def volume=(value)
-    unless (0..1).cover?(value)
-      raise ArgumentError, "Value must fall between 0 and 1, you gave me #{value}"
-    end
-    @volume = value
-
-    set_music_volume(self, value)
-  end
-
   # Set the pitch.
   # @param value [Float]
   # @return [nil]